azure_function/requirements.txt
azure-functions
openai>=1.12.0
pyodbc>=5.0.1
python-dotenv>=1.0.0
